+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to BrowserUtils will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2026-06-27
+
+### Added
+- Paste Anyway can now allow pasting into text fields on sites that block paste events
+- Settings now include a Paste Anyway toggle that applies immediately after changing it
+
+### Changed
+- Settings controls now autosave on change instead of requiring repeated Save Settings buttons
+- Feature releases use the middle version number, while patch-only fixes use the right version number
+
 ## [0.10.19] - 2026-06-22
 
 ### Changed